requirements.txt
Flask
python-dotenv
s3fs
PyPDF2
pandas
doctr
torch
thefuzz
